requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
openai==1.61.1
python-dotenv==1.0.1
pandas==2.2.3
kagglehub==0.3.8

